Protect against bad channel data.
This adds two protections.  First when a bad channel trailing word is
found, this continues to process the event (doesn't throw a runtime error).
Second, this makes sure that the output digit is never longer than 9596
(the maximum integration window).  It might be better to put the length
check outside of the translation routines, but, for now, this protects the
rest of the software from extra long data.